Only open image once
